pyproject.toml
[project]
name = "pochitrain"
version = "1.4.3"
description = "Simple CNN training framework for image classification"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "kurorosu" }
]

dependencies = [
    "torch",
    "torchvision",
    "numpy==1.26.1; python_version < '3.11'",
    "numpy>=2.4.1; python_version >= '3.11'",
    "pandas>=2.0.0",
    "pillow>=8.0.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.10.8",
    "pyyaml>=5.4.0",
    "colorlog>=6.0.0",
    "rich>=13.0.0",
    "optuna>=3.5.0",
    "plotly>=5.0.0",
    "matplotlib-fontja>=1.1.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pydocstyle>=6.3.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
]
onnx = [
    "onnx>=1.14.0",
    "onnxscript>=0.1.0",
    "onnxruntime-gpu>=1.16.0",
]

[project.scripts]
pochi = "pochitrain.cli.pochi:main"
export-onnx = "pochitrain.cli.export_onnx:main"
infer-onnx = "pochitrain.cli.infer_onnx:main"
infer-trt = "pochitrain.cli.infer_trt:main"
vis-grad = "pochitrain.cli.visualize_gradient:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pydocstyle]
convention = "google"
add-ignore = ["D212"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -n 6 --dist=worksteal"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
